requirements.txt
absl-py==1.4.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
filelock==3.9.0
fonttools==4.40.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.0
json-tricks==3.17.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.3.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.3.6
yacs==0.1.8
zipp==3.15.0
